Group DTOs by their model in single files
