--- Core/Dtos/University/ActivityTypeMinimalDto.cs
@@ -1,3 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record ActivityTypeMinimalDto(int Id, string Name);
