--- Core/Dtos/University/ActivityTypeViewDto.cs
@@ -1,4 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record ActivityTypeViewDto(int Id, string Name,
-    DateTimeOffset CreationDate, DateTimeOffset UpdateDate);
