--- Core/Dtos/University/ClassTypeViewDto.cs
@@ -1,4 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record ClassTypeViewDto(int Id, string Name,
-    DateTimeOffset CreationDate, DateTimeOffset UpdateDate);
