--- Core/Dtos/University/ClassUpdateDto.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
-
-namespace EUniversity.Core.Dtos.University;
-
-[ValidateNever] // Remove data annotations validation
-public record ClassUpdateDto(int ClassTypeId,
-    int ClassroomId, int GroupId, string? SubstituteTeacherId,
-    DateTimeOffset StartDate, TimeSpan Duration) : IClassWriteDto;
