+++ Core/Dtos/University/ClassroomDtos.cs
@@ -4,3 +4,7 @@ namespace EUniversity.Core.Dtos.University;
 
 [ValidateNever] // Remove data annotations validation
 public record ClassroomCreateDto(string Name);
+
+public record ClassroomViewDto(int Id, string Name,
+    DateTimeOffset CreationDate, DateTimeOffset UpdateDate);
+
