--- Core/Dtos/University/ClassroomViewDto.cs
@@ -1,4 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record ClassroomViewDto(int Id, string Name,
-    DateTimeOffset CreationDate, DateTimeOffset UpdateDate);
