--- Core/Dtos/University/CourseMinimalViewDto.cs
@@ -1,3 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record CourseMinimalViewDto(int Id, string Name, SemesterMinimalViewDto? Semester);
