--- Core/Dtos/University/CoursePreviewDto.cs
@@ -1,5 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record CoursePreviewDto(int Id, string Name,
-    DateTimeOffset CreationDate, DateTimeOffset UpdateDate,
-    SemesterMinimalViewDto? Semester);
