--- Core/Dtos/University/Grades/AssignedGradeUpdateDto.cs
@@ -1,6 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
-
-namespace EUniversity.Core.Dtos.University.Grades;
-
-[ValidateNever] // Remove data annotations validation
-public record class AssignedGradeUpdateDto(int GradeId, string? Notes, int? ActivityTypeId);
