--- Core/Dtos/University/Grades/AssignedGradeViewDto.cs
@@ -1,10 +1,0 @@
-﻿using EUniversity.Core.Dtos.Users;
-
-namespace EUniversity.Core.Dtos.University.Grades;
-
-public record AssignedGradeViewDto(int Id, string? Notes,
-    GradeMinimalViewDto Grade, GroupMinimalViewDto? Group,
-    UserViewDto? Assigner, UserViewDto? Reassigner,
-    StudentPreviewDto? Student,
-    DateTimeOffset CreationDate, DateTimeOffset UpdateDate,
-    ActivityTypeMinimalDto? ActivityType);
