--- Core/Dtos/University/Grades/GradeMinimalViewDto.cs
@@ -1,3 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University.Grades;
-
-public record GradeMinimalViewDto(int Id, string Name);
