--- Core/Dtos/University/Grades/GradeViewDto.cs
@@ -1,4 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University.Grades;
-
-public record GradeViewDto(int Id, string Name, int Score,
-    DateTimeOffset CreationDate, DateTimeOffset UpdateDate);
