--- Core/Dtos/University/GroupMinimalViewDto.cs
@@ -1,3 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record GroupMinimalViewDto(int Id, string Name, CourseMinimalViewDto Course);
