--- Core/Dtos/University/GroupPreviewDto.cs
@@ -1,5 +1,0 @@
-﻿using EUniversity.Core.Dtos.Users;
-
-namespace EUniversity.Core.Dtos.University;
-
-public record GroupPreviewDto(int Id, string Name, DateTimeOffset CreationDate, DateTimeOffset UpdateDate, TeacherPreviewDto? Teacher, CoursePreviewDto Course);
