--- Core/Dtos/University/SemesterMinimalViewDto.cs
@@ -1,3 +1,0 @@
-﻿namespace EUniversity.Core.Dtos.University;
-
-public record SemesterMinimalViewDto(int Id, string Name);
